change schemas of preferred author

+++ src/main/kotlin/com/wutsi/blog/client/view/PreferredAuthorDto.kt
@@ -2,5 +2,5 @@ package com.wutsi.blog.client.view
 
 data class PreferredAuthorDto(
         val authorId: Long = -1,
-        val score: Long = 0
+        val score: Double = 0.0
 )
